refactor: extract useEndingDetection hook from NarrativeController
NarrativeController.tsx was 1,364 lines with three callbacks over 200
lines each. Extract checkForEndingIndicators (the 109-line pure-AI ending
analysis function) into a useEndingDetection hook that also owns the
endingSuggestedRef "fire once per session" gate.

The hook exposes:
- checkForEndingIndicators(segment): runs AI analysis, fires
  onEndingSuggested for medium/high confidence responses
- suggestEnding(reason, type): atomic "if not already suggested, fire" —
  consumed by the two short-circuit paths in NarrativeController
  (critical-decision failure, fatal-outcome tag) that previously each
  duplicated the check-then-set-ref-then-call dance.

Pulls a handful of magic numbers (RECENT_SEGMENT_WINDOW=5,
LONG_STORY_THRESHOLD=10, EARLIER_CONTEXT_TRUNCATE_CHARS=500,
MIN_SEGMENTS_FOR_ANALYSIS=3) into named constants.

NarrativeController.tsx: 1,364 -> 1,212 lines. First foothold; the two
remaining mega-callbacks (generatePlayerChoices ~349 lines and
generateNextSegment ~359 lines) still need extracting.
